requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
line-bot-sdk==3.21.0
google-genai==0.2.2
python-dotenv==1.0.1
pydantic-settings==2.7.0
aiosqlite==0.20.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0